# Conflicts:
#	app/build.gradle
#	app/src/main/java/com/example/jeroen_van_ottelen/ikpmd_nieuwe_app/MainActivity.java
#	app/src/main/java/com/example/jeroen_van_ottelen/ikpmd_nieuwe_app/activity/MainActivity.java
#	app/src/main/java/com/example/jeroen_van_ottelen/ikpmd_nieuwe_app/activity/OverzichtActivity.java
#	app/src/main/java/com/example/jeroen_van_ottelen/ikpmd_nieuwe_app/database/DatabaseReceiver.java
#	app/src/main/java/com/example/jeroen_van_ottelen/ikpmd_nieuwe_app/models/Subject.java
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/jeroen_van_ottelen/ikpmd_nieuwe_app/models/Subject.java
@@ -9,7 +9,7 @@ public class Subject {
 	private String name;
 	private int ects;
 	private int period;
-	private int grade;
+	private double grade;
 
 	public String getName()
 	{
@@ -41,7 +41,7 @@ public class Subject {
 		this.period = period;
 	}
 
-	public int getGrade()
+	public double getGrade()
 	{
 		return grade;
 	}
